Backend/requirements.txt
# FastAPI dependencies
fastapi[all]==0.104.1
uvicorn[standard]==0.24.0

# Database
sqlalchemy==2.0.35
psycopg[binary]==3.2.9
alembic==1.13.3

# Authentication & Security
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.6
firebase-admin==6.2.0
google-auth==2.23.4

# Configuration
python-decouple==3.8
pydantic-settings==2.0.3

# Email validation
email-validator==2.1.0

# Development tools
pytest==7.4.3
pytest-asyncio==0.21.1
httpx==0.25.2

# Email sending
aiosmtplib==3.0.1
jinja2==3.1.2

# Date/time handling
python-dateutil==2.8.2

# Logging
loguru==0.7.2
